BackEnd/requirements.txt
alembic==1.14.0
attrs==24.2.0
bcrypt==4.2.1
blinker==1.9.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.2
flasgger==0.9.7.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.6
MarkupSafe==3.0.2
mistune==3.0.2
PyJWT==2.10.0
python-dotenv==1.0.1
PyYAML==6.0.2
six==1.16.0
SQLAlchemy==2.0.36
tomli==2.1.0
typing_extensions==4.12.2
Werkzeug==3.1.3
zipp==3.21.0
